chore: Move all "dynamicLoaders" into composite
Remove loader from util as it was never fully implemented.
Cleaned up ddl-shim and updated to latest (removed ajv dynamic loader).
Moved Persist clone and create into composite.
