+++ packages/composite/src/index.ts
@@ -2,3 +2,7 @@ export * from "./ChartPanel";
 export * from "./Dermatology";
 export * from "./MegaChart";
 export * from "./MultiChart";
+import * as Persist from "./Persist";
+export { Persist };
+import * as Utility from "./Utility";
+export { Utility };
