+++ packages/ddl-shim/src/index.ts
@@ -1,10 +1,14 @@
-import * as Ajv from "ajv";
-
 import * as DDL1 from "./ddl";
 import { ddlSchema } from "./ddlSchema";
 
 import * as DDL2 from "./ddl2";
 import { ddl2Schema } from "./ddl2Schema";
+
+// @ts-ignore
+import validateSchema from "../src/ddlSchema.json";
+
+// @ts-ignore
+import validateSchema2 from "../src/ddl2Schema.json";
 
 export { DDL1, DDL2, ddlSchema, ddl2Schema };
 export * from "./upgrade";
@@ -15,17 +19,15 @@ export interface Response {
 }
 
 export function validate(ddl: object): Response {
-    const ajv = new Ajv();
     return {
-        success: ajv.validate(ddlSchema, ddl),
-        errors: ajv.errors
+        success: validateSchema(ddl),
+        errors: validateSchema2.errors
     };
 }
 
 export function validate2(ddl: object): Response {
-    const ajv = new Ajv();
     return {
-        success: ajv.validate(ddl2Schema, ddl),
-        errors: ajv.errors
+        success: validateSchema2(ddl),
+        errors: validateSchema2.errors
     };
 }
