+++ packages/util/src/index.ts
@@ -5,7 +5,6 @@ export * from "./dictionary";
 export * from "./esp";
 export * from "./graph";
 export * from "./hashSum";
-export * from "./loader";
 export * from "./logging";
 export * from "./object";
 export * from "./observer";
